requirements.txt
easyocr
opencv-python-headless
matplotlib
pillow
streamlit
pandas
numpy
streamlit-webrtc
